I used double images.
Thanks for contributing many images, Merts!

valorant-v8.pt and v9 are almost equal. 
Why is it almost equal? Because Google Colab wasn't able to process due to a lot of images. 
So I changed batch-size 80 to 50. It turned out to be almost the same.

If anyone knows how to process 4126 Train Images and 2675 Val Images in Google Colab Pro with batch-size 80, Please let me know.